requirements.txt
boto3
mysql-connector-python
dotenv
openai